packages/engine/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
annotated-types==0.7.0
anthropic==0.71.0
anyio==4.11.0
asyncio-mqtt==0.16.2
attrs==25.4.0
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
crcmod==1.7
cryptography==46.0.3
dashscope==1.24.7
distro==1.9.0
docstring_parser==0.17.0
frozenlist==1.8.0
google-ai-generativelanguage==0.6.15
google-api-core==2.27.0
google-api-python-client==2.185.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.71.0
greenlet==3.2.4
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
Jinja2==3.1.6
jiter==0.11.1
jmespath==0.10.0
jsonpath-ng==1.7.0
loguru==0.7.3
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
midscene-python==0.1.1
multidict==6.7.0
networkx==3.5
numpy==1.26.4
openai==1.109.1
opencv-python==4.11.0.86
oss2==2.18.0
outcome==1.3.0.post0
paho-mqtt==2.1.0
pillow==10.4.0
playwright==1.55.0
ply==3.11
propcache==0.4.1
proto-plus==1.26.1
protobuf==5.29.5
pure-python-adb==0.3.0.dev0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.11.10
pydantic_core==2.33.2
pyee==13.0.0
Pygments==2.19.2
PyMySQL==1.1.2
pyparsing==3.2.5
PySocks==1.7.1
PyYAML==6.0.3
requests==2.32.0
rich==14.2.0
rsa==4.9.1
selenium==4.38.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
tqdm==4.67.1
trio==0.31.0
trio-websocket==0.12.2
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
websocket-client==1.9.0
websockets==15.0.1
wsproto==1.2.0
yarl==1.22.0
